#iss2 Edited code to clean it up. Edited the constructor to take Strings for the sender and recipient.

+++ Pavle/CompareMessageByReadAndTime.java
@@ -4,9 +4,8 @@ public class CompareMessageByReadAndTime implements Comparator<Message> {
 
    @Override
    public int compare(Message m1, Message m2) {
-      int read_status = m1.read - m2.read;
+      int read_status = (int)(m1.read - m2.read);
       if (read_status == 0) return -Long.signum(m1.timeSent - m2.timeSent);
-      // -m1.timeSent.compareTo(m2.timeSent);
       return read_status;
    }
